spending/pyproject.toml
[project]
name = "spending"
python = "3.13.5"
version = "0.1.0"
description = ""
authors = [
    {name = "ku113p@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp (>=3.12.13,<4.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pymongo (>=4.13.2,<5.0.0)",
    "langgraph (>=0.5.1,<0.6.0)",
    "langchain[openai] (>=0.3.26,<0.4.0)",
    "langfuse (>=3.1.3,<4.0.0)",
    "langchain-openai (>=0.3.27,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "redis (>=6.2.0,<7.0.0)",
    "langgraph-checkpoint-redis (>=0.0.8,<0.0.9)",
    "python-telegram-bot (>=22.3,<23.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "3.13.5"

[tool.poetry.group.dev.dependencies]
notebook = "^7.4.4"
ipykernel = "^6.29.5"
python-dotenv = {extras = ["cli"], version = "^1.1.1"}

